Move SignData to main project
