+++ LegacySignController/src/SignData/SignConfigurationData.cpp
