+++ LegacySignController/src/SignData/SignConfigurationData.h
