--- lib/SignDataLib/SignDataLib.h
@@ -1,8 +1,0 @@
-#ifndef SIGNDATALIB_H
-#define SIGNDATALIB_H
-
-#include "SignOffsetData.h"
-#include "SignStatus.h"
-#include "SignConfigurationData.h"
-
-#endif
